+++ CHANGELOG.md
@@ -1,31 +1,45 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
 # zentao-plugin Changelog
+
+## [1.0.6]
+
+### Fixed
+
+- refactor token refresh logic
+
 ## [1.0.5]
+
 ### Features
+
 - Support for IntelliJ 2025.3
 
-
 ## [1.0.4]
+
 ### Features
+
 - Support for IntelliJ 2025.1
 
-
 ## [1.0.3]
+
 ### Features
+
 - Support for IntelliJ 2024.3
 
-
 ## [1.0.2]
+
 ### Features
+
 - Support for IntelliJ 2024.2
 
-
 ## [1.0.1]
+
 ### Features
+
 - Support for IntelliJ 2024.1
 
-
 ## [1.0.0]
+
 ### Features
-- Add zentao support to IntelliJ Task Management 
+
+- Add zentao support to IntelliJ Task Management